+++ e2e/__tests__/globalTeardown.test.ts
@@ -40,7 +40,7 @@ test('globalTeardown is triggered once after all test suites', () => {
   const teardownPath = path.resolve(e2eDir, 'teardown.js');
   const result = runWithJson('global-teardown', [
     `--globalTeardown=${teardownPath}`,
-    '--testPathPattern=__tests__',
+    '--testPathPatterns=__tests__',
   ]);
 
   expect(result.exitCode).toBe(0);
@@ -54,7 +54,7 @@ test('jest throws an error when globalTeardown does not export a function', () =
   const teardownPath = path.resolve(e2eDir, 'invalidTeardown.js');
   const {exitCode, stderr} = runJest(e2eDir, [
     `--globalTeardown=${teardownPath}`,
-    '--testPathPattern=__tests__',
+    '--testPathPatterns=__tests__',
   ]);
 
   expect(exitCode).toBe(1);
@@ -67,15 +67,13 @@ test('jest throws an error when globalTeardown does not export a function', () =
 test('globalSetup function gets global config object and project config as parameters', () => {
   const teardownPath = path.resolve(e2eDir, 'teardownWithConfig.js');
 
-  const testPathPattern = 'pass';
-
   const result = runJest(e2eDir, [
     `--globalTeardown=${teardownPath}`,
-    `--testPathPattern=${testPathPattern}`,
+    '--testPathPatterns=pass',
     '--cache=true',
   ]);
 
-  expect(result.stdout).toBe(`${testPathPattern}\ntrue`);
+  expect(result.stdout).toBe("[ 'pass' ]\ntrue");
 });
 
 test('should call globalTeardown function of multiple projects', () => {
@@ -95,7 +93,7 @@ test('should not call a globalTeardown of a project if there are no tests to run
 
   const result = runWithJson('global-teardown', [
     `--config=${configPath}`,
-    '--testPathPattern=project-1',
+    '--testPathPatterns=project-1',
   ]);
 
   expect(result.exitCode).toBe(0);
@@ -108,15 +106,13 @@ test('should not call a globalTeardown of a project if there are no tests to run
 test('globalTeardown works with default export', () => {
   const teardownPath = path.resolve(e2eDir, 'teardownWithDefaultExport.js');
 
-  const testPathPattern = 'pass';
-
   const result = runJest(e2eDir, [
     `--globalTeardown=${teardownPath}`,
-    `--testPathPattern=${testPathPattern}`,
+    '--testPathPatterns=pass',
     '--cache=true',
   ]);
 
-  expect(result.stdout).toBe(`${testPathPattern}\ntrue`);
+  expect(result.stdout).toBe("[ 'pass' ]\ntrue");
 });
 
 test('globalTeardown throws with named export', () => {
@@ -127,7 +123,7 @@ test('globalTeardown throws with named export', () => {
 
   const {exitCode, stderr} = runJest(e2eDir, [
     `--globalTeardown=${teardownPath}`,
-    '--testPathPattern=__tests__',
+    '--testPathPatterns=__tests__',
   ]);
 
   expect(exitCode).toBe(1);