+++ packages/jest-core/src/plugins/FailedTestsInteractive.ts
@@ -58,7 +58,7 @@ export default class FailedTestsInteractivePlugin extends BaseWatchPlugin {
         updateConfigAndRun({
           mode: 'watch',
           testNamePattern: failure ? `^${failure.fullName}$` : '',
-          testPathPattern: failure?.path || '',
+          testPathPatterns: failure ? [failure.path] : [],
         });
 
         if (!this._manager.isActive()) {