+++ packages/jest-core/src/watch.ts
@@ -15,6 +15,7 @@ import type {Config} from '@jest/types';
 import type {IHasteMap as HasteMap} from 'jest-haste-map';
 import {formatExecError} from 'jest-message-util';
 import {
+  TestPathPatterns,
   isInteractive,
   preRunMessage,
   requireOrImportModule,
@@ -120,7 +121,7 @@ export default async function watch(
     onlyFailures,
     reporters,
     testNamePattern,
-    testPathPattern,
+    testPathPatterns,
     updateSnapshot,
     verbose,
   }: AllowedConfigOptions = {}) => {
@@ -140,7 +141,7 @@ export default async function watch(
       onlyFailures,
       reporters,
       testNamePattern,
-      testPathPattern,
+      testPathPatterns,
       updateSnapshot,
       verbose,
     });
@@ -227,9 +228,12 @@ export default async function watch(
 
   const emitFileChange = () => {
     if (hooks.isUsed('onFileChange')) {
+      const testPathPatterns = new TestPathPatterns([], globalConfig);
       const projects = searchSources.map(({context, searchSource}) => ({
         config: context.config,
-        testPaths: searchSource.findMatchingTests('').tests.map(t => t.path),
+        testPaths: searchSource
+          .findMatchingTests(testPathPatterns)
+          .tests.map(t => t.path),
       }));
       hooks.getEmitter().onFileChange({projects});
     }
@@ -404,7 +408,7 @@ export default async function watch(
         globalConfig = updateGlobalConfig(globalConfig, {
           mode: 'watchAll',
           testNamePattern: '',
-          testPathPattern: '',
+          testPathPatterns: [],
         });
         startRun(globalConfig);
         break;
@@ -412,7 +416,7 @@ export default async function watch(
         updateConfigAndRun({
           mode: 'watch',
           testNamePattern: '',
-          testPathPattern: '',
+          testPathPatterns: [],
         });
         break;
       case 'f':
@@ -425,7 +429,7 @@ export default async function watch(
         globalConfig = updateGlobalConfig(globalConfig, {
           mode: 'watch',
           testNamePattern: '',
-          testPathPattern: '',
+          testPathPatterns: [],
         });
         startRun(globalConfig);
         break;
@@ -528,10 +532,11 @@ const usage = (
   watchPlugins: Array<WatchPlugin>,
   delimiter = '\n',
 ) => {
+  const testPathPatterns = TestPathPatterns.fromGlobalConfig(globalConfig);
   const messages = [
     activeFilters(globalConfig),
 
-    globalConfig.testPathPattern || globalConfig.testNamePattern
+    testPathPatterns.isSet() || globalConfig.testNamePattern
       ? `${chalk.dim(' \u203A Press ')}c${chalk.dim(' to clear filters.')}`
       : null,
     `\n${chalk.bold('Watch Usage')}`,
@@ -549,7 +554,7 @@ const usage = (
         )}`,
 
     (globalConfig.watchAll ||
-      globalConfig.testPathPattern ||
+      testPathPatterns.isSet() ||
       globalConfig.testNamePattern) &&
     !globalConfig.noSCM
       ? `${chalk.dim(' \u203A Press ')}o${chalk.dim(