+++ packages/jest-watcher/src/types.ts
@@ -62,7 +62,7 @@ export type AllowedConfigOptions = Partial<
     | 'onlyFailures'
     | 'reporters'
     | 'testNamePattern'
-    | 'testPathPattern'
+    | 'testPathPatterns'
     | 'updateSnapshot'
     | 'verbose'
   > & {mode: 'watch' | 'watchAll'}